nvme: fio_plugin PRACT=1 incorrect
The NVMe FIO plugin was written to support PRACT=1 but the
implementation was incomplete.

PRACT=1 is a bit within read/write I/O commands which causes
the NVMe device to NOT transfer any metadata between the
host and the controller.

As the present implementation included an option to set
PRACT (pi_act), setting it caused commands to have PRCHK and
PRACT fields clear, meaning metadata was transferred but not
checked.

This mod "finishes" PRACT support, so that commands are
properly formed - only user data are transferred and the
PRCHK and SLBA, App tag, and App mask within the NVMe
command block are now correct.

Signed-off-by: James Bergsten <jamesx.bergsten@intel.com>
Change-Id: I930c8a07519a4742c44240801b068fac2c4802a7