Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

Thank you for adding a test to make sure this functionality works!

My only concern is that a year from now, it may be confusing for someone to look at this code and understand why the script is using the alias RPC.  Could we create a completely separate test for this include_aliases functionality?  It could be as simple as trying to load_config with a single construct_malloc_bdev RPC function.  This will make the test much more explicit and not buried in the vhost tests.

Patch-set: 1
Reviewer: Gerrit User 1011222 <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1