Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> (2 comments)
> 
> Hi Paul,
> 
> I think I will be wrong but I couldn't understand how src_mbufs[src_iov_total] is used. Will you teach me?
> 
> -Shuhei

Sure, it's a great question.  We have to allocate exactly the number of mbufs we need and they need to be passed to DPDK in an array, like the one you ask about.  When we need an extra one because of the 2MB boundary thing, which we don't know up front, we have to allocate another mbuf into the last position in the array so that's what src_iov_total keeps track of.  It's also used to make sure we free all allotted mbufs in the error path too.

Patch-set: 6
Reviewer: Gerrit User 1011289 <1011289@d5d70762-12d0-45a1-890d-524b12d3f735>