Update patch set 7
Patch Set 7:

BTW, the reason there is not a UT with this to prove the question you ask is a very bad answer - they have been broken for both reduce and compress for quite some time but because of a mistake in CI were not being ran so nobody knew.  I discovered this when I did this patch which is critical for data integrity using QAT so I didn't want to hold it up to fix the existing broken tests and then add new ones.  I'll prioritize this for tomorrow though.

Patch-set: 7
Reviewer: Gerrit User 1011289 <1011289@d5d70762-12d0-45a1-890d-524b12d3f735>