Update patch set 7
Patch Set 7:

Shuhei's observation is correct, if the IOV that crosses a boundary is not the last in the list (which is has been in practice) here the current code will reuse the same mbuf array position on the subsequent mbuf thus corruption the chain.  I don't believe the suggested fix is completely correct though.

What's here certainly fixes CI tests though so is better than what was before, feel free to merge but I should be able to have a solid fix for this later today and if we want to wait for good UT to go with it that might be a few days since they're broken right now.

Patch-set: 7
Reviewer: Gerrit User 1011289 <1011289@d5d70762-12d0-45a1-890d-524b12d3f735>