Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3: Code-Review-1
> > 
> > After looking at following patches in series, I think there is a better way to handle errors using the interface. Mode parameter seems to complicate the interface for the user and we should be able to achieve the same using only the rollback argument
> > 
> > It should work like this:
> > - only mngt_stop() stops the execution, not mngt_continue()
> > - rollback argument is a part of *_path table (which is null terminated array of vbdev_ocf_mngt_fn)
> > - if rollback is not null, mngt_stop() cleans vbdev_ocf_mngt_ctx structure and sets vbdev->mngt_ctx->current_step = rollback and calls rollback
> > - if rollback is null, mngt_stop exits normally
> > - errors produced by one of the steps are handled by calling mngt_stop() immediately
> > - timeout error is handled by checking return status of previus function and calling mngt_stop() (extra step can be added to handle all errors of previus one)
> > 
> > So this patch can be abandoned if you agree with proposed alternative solution
> 
> From what I see, you want to remove mode to simplify interface but also you are moving responsibility of checking and handling (also from inside of interface) in each step for the future user. I'm not sure that is better way, but I'm not personally connected to this changes, so if you want to push patches with your approach I'm ok with abandoning my series.

Yes, I want to keep interface as simple as possible. Mode parameter compicates it because it creates 2 completely different flows.
Also, having mngt_continue() to stop execution is not intuitive.
You are right that in my approach user gets responsibility for checking and handling each error, but that is good because it's more explicit which is good for reader and it also gives more control for the user.
Mode parameter seems really good for handling timeout error, but not errors that happen in usual steps IMO. Maybe we should have simmilar parameter specifically in mngt_poll() function.
Another related aspect is that my approach will make rollback much easier to implement - you don't need to make copies and helper functions.
Your approach is nice and solves the error handling, but we also want to have code that is as much readable and maintainable as possible, so let's discuss and develop solution that we all agree on.

Patch-set: 3
Reviewer: Gerrit User 1014935 <1014935@d5d70762-12d0-45a1-890d-524b12d3f735>