Update patch set 3
Patch Set 3: Code-Review-1

I'll provide a bit of background that's missing here, which should hopefully make everything a lot clearer. I'll start by saying that the connection scheduling code is intentionally pushed up into the application - that's a very carefully thought out design choice.

In order for SPDK to be integrated into a wide variety of applications and frameworks, we need to keep the "framework" out of the SPDK libraries. This mostly means that SPDK libraries need to be designed to integrate into some other application's threading model, as much as possible. SPDK certainly has to assume there is some amount of cooperative multi-tasking going on, but it tries to make as few other assumptions as possible. We always say our real deliverables are the things in /lib, not the things in /app.

This is why you see the application (lib/event/subsystems stuff) creating the poll groups and assigning qpairs to them - placing poll groups on threads and loading them up with work is really an application-level decision and I didn't want the library making decisions like that. As the code works today, the connection scheduling needs to stay in the application code.

However, as we transition to our new lightweight threading model the situation improves. This allows SPDK libraries to spawn their own lightweight threads and whatever application is in use then schedules them onto the real cores. When we fully arrive at this new model, we can simply have lib/nvmf spawn some number of lightweight threads, put a poll group on each, and do connection scheduling internally. The user can still load balance by now moving the lightweight threads around amongst the system threads, as opposed to moving qpairs between poll groups. So, we're getting there and heading in a direction that aligns with your goals in this patch series, but we're not there yet.

Patch-set: 3
Reviewer: Gerrit User 1010531 <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1