Update patch set 2
Patch Set 2: Code-Review-1

(3 comments)

Hi Eugene,

Thanks for catching this. It looks like I missed a condition on line 661. The original mechanism that was supposed to catch the crossover between requests can be found in line 668. We were branching based off of the wr_id which is set to point at the data object of the request. This is a fool proof way of catching that we don't modify any objects not related to the request, but I foolishly missed the check on line 661 when updating the function.

I am personally partial to keeping the check based off of the wr_id so that we don't have to add another field that we need to assign and remove. Of course using the wr_id method means a couple of extra checks per loop.
The -1 is just for comment visibility. If everyone agrees that adding the new tracker variable is better than using the old mechanism, I am fine to merge this as is.
Either way, I think we should keep your loop simplification and the memset which both clean up the code considerably.

Patch-set: 2
Reviewer: Gerrit User 1011223 <1011223@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1