check_format: don't enforce single quotes in bash traps
There's really no reason to enforce either single or
double quotes in traps. They have different behavior,
but it's not like one is better than the other.

Single quote trap definitions are evaluated when the
traps are trigerred, double quotes are evaluated at
the time of setting up the trap.

Change-Id: I577a6d50168ffe7a81e643ff32bb7bf32bd1955c
Signed-off-by: Darek Stojaczyk <dariusz.stojaczyk@intel.com>