bdev/compress: small fix-up to 2MB patch
For completion context, the first src mbuf 'userdata' field
is filled in. The wrong index was being used to set this which
would never cause a problem because only the first src mbuf is
used.  Also when an extra src mbuf is added we weren't setting
userdata which again isn't a real problem but for consistency
I'd like to set it since all other src mbufs have it set.

Change-Id: Iac32a9a37502a95ce0f997375b6c75a42bc1651f
Signed-off-by: paul luse <paul.e.luse@intel.com>