Update patch set 11
Patch Set 11:

> Patch Set 11:
> 
> Hi Ziye,
> 
> My comment is not for this patch yet.
> 
> You are proposing another idea to insert delay to every full socket read. If I understand correctly, the another idea has better performance improvement, more than 10%. These two ideas are additive, i.e., can we get 10 + 5 = 15% improvement?
> 
> Another idea is much simpler than this patch and will not cause fair share issue among connections. I also saw the same improvement. But our concern is that the parameter tuning is very sensitive and difficult to find, and we don't know exactly why the another idea works.
> 
> Did you abandon the another idea? Will you share your strategy?, and how do you think about stabilizing spdk + vpp rather than tuning spdk + kernel?
> 
> Thanks,
> Shuhei

Hi Shuhei,

Currently, we hope that we can get the tradeoff between the system call and memory copy. So the strategy is that:

1 For the read/recv system call, we should use less system calls to reduce the context switch between kernel and user space.
2 Though we reduced the system calls, however there is memory copy, so we need to seek a trade off.
3 We may use async read/recv/writev later (still need to investigate).

Currently, if we use epoll_wait, I think that my previous patch to reduce the epoll_wait should be still valid, though Jim/Ben do not want to check it in.

For VPP + DPDK + NVMe-oF, I think that it should not be a good direction from my personal mind. I do not against user space solution, but I think VPP is not a suitable candidate, until VPP can directly be used as lib to run with SPDK together instead of a standalone process.

Patch-set: 11
Reviewer: Gerrit User 1011275 <1011275@d5d70762-12d0-45a1-890d-524b12d3f735>