Makefile: don't override MAKEFLAGS in submake
Apparently make doesn't like it:
make[2]: warning: -jN forced in submake: disabling jobserver mode.

We only did it because of scan-build - it detects a ton
of issues (or false positives) in our dependencies. To
disable it, we don't have to override MAKEFLAGS, but
just CC - that's what we'll do now.

Fixes #896

Change-Id: I5eea984d6bbfbf4caabdd704850fac840fed3524
Signed-off-by: Darek Stojaczyk <dariusz.stojaczyk@intel.com>