nvmf_example: add an nvmf example
use the nvmf lib to add an nvmf example which
doesn't use the spdk app framework. It can show
how to use the nvmf lib to build your own nvmf target.

setup the SPDK env and parse the args. Use the
conf lib to make it easier to realize or you
can use RPC cmd. Default there is no configuration
file. We just add a peremeter and the RPC relization
is in the next patch .

Change-Id: I8d175934f7364d40fb44d36e85c31d8ea5063654
Signed-off-by: JinYu <jin.yu@intel.com>