Update patch set 12
Patch Set 12: Code-Review+2

Hi Ziye,

Thanks to your lecture, I understood that big buffer is per queue pair and we choose if we use it or not according to the destination. The amount of inserted delay was 750ms and pretty big in my environment. Latency of copy will be negligible compared with that.

-Shuhei

Patch-set: 12
Reviewer: Gerrit User 1012251 <1012251@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+2