autobuild: don't build the tests under scan-build.
Scan-build has been throwing an ever-increasing number of false
positives against the unit tests. For that reason, we make scan-build
without making the tests so that we only get relevant errors from the
apps, libraries, modules, and examples. Any test applications or
unittests will not be tested. However, they will still be run under
ASAN when that flag is enabled.

Change-Id: I00b0f0ed4d10dc3623e049634ca64926fb12592d
Signed-off-by: Seth Howell <seth.howell@intel.com>