Update patch set 2
Patch Set 2:

(1 comment)

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > (1 comment)
> 
> I would like to leave the default number of queues as it is because it may expose more issues but not make it hidden.  As for the iteration number, I don't think 10 or 200 make too much difference.

Hi Changpeng, one reason we want to perform connect_disconnect a lot of times is to make sure that qpair resources are being freed back to the system. In the context of how many resources might be consumed by defunct connections, there is a big difference between 10 and 200. From that perspective, we should probably even decrease the amount of memory allocated to SPDK and increase the number of times we connect and disconnect qpairs.

Patch-set: 2
Reviewer: Gerrit User 1011223 <1011223@d5d70762-12d0-45a1-890d-524b12d3f735>