nvmf: Hold requested number of buffers into struct spdk_nvmf_request
This is a preparation to make spdk_nvmf_request_get_buffers()
asynchronous. When spdk_nvmf_request_get_buffers() is asynchronous,
it will be retried later if it fails due to lack of memory.

For that case, the requested number of buffers is saved into struct
spdk_nvmf_request and used on retry.

As a side effect, we can remove num_buffers from
spdk_nvmf_request_free_buffers() because it is already saved in
req->num_buffers.

Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>
Change-Id: Ifafb246a6d5ee1692c687e6b5e0c83874b4af3c2