nvmf/tcp: Move state transition to just after completion of buffer allocation
This is a preparation to make I/O buffer allocation asynchronous.

As a result of this change, two debug logs became almost duplicated
We will see subtle difference but they are debug logs and will be
replaced eventually. So to simplify the code, remove the later one
in this patch.

Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>
Change-Id: I502b995562774e04bad73b67f37386b4218bfa06