Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > (1 comment)
> > 
> > > Patch Set 2:
> > > 
> > > > Patch Set 2:
> > > > 
> > > > (1 comment)
> > > 
> > > I would like to leave the default number of queues as it is because it may expose more issues but not make it hidden.  As for the iteration number, I don't think 10 or 200 make too much difference.
> > 
> > Hi Changpeng, one reason we want to perform connect_disconnect a lot of times is to make sure that qpair resources are being freed back to the system. In the context of how many resources might be consumed by defunct connections, there is a big difference between 10 and 200. From that perspective, we should probably even decrease the amount of memory allocated to SPDK and increase the number of times we connect and disconnect qpairs.
> 
> 
> You may see we have several github issues related with limited RDMA resources, we can limit the number of queues as a workaround, but we didn't do that, just upgrade the kernel version. I think for this case you can add a new test case just allocate 1 IO queue and just do many times connect/disconnect tests.

I don't really see a need for a new test case. We would only lower the number of queues for this particular script, others would still use default values so we still would we able to detect issues with higher IO queue numbers as you mentioned before.

Here's an idea though - what if we limit the number of created IO queues for nightly test only?
So for 10 connections (in per-patch tests) we'd use anything which nvme-cli uses as default value, but for nightly we'd limit it to 8.

Patch-set: 2
Reviewer: Gerrit User 1011256 <1011256@d5d70762-12d0-45a1-890d-524b12d3f735>