lib/blobstore: respect clear_method when deleting an lvol
When detroying an lvol, use the proper clear method as was
specified when it was created. This requires use of a new
spdk_bs_delete_blob_ext() function that can accept the
clear_method parameter when spdk_lvol_destroy() is called
so that the blob can be opened for deletion with
spdk_bs_open_blob_ext() in order to pass the opts struct.

Fixes issue #944

Change-Id: I8f78e2fa25a602cac44c5a9182aafc04e728ccd3
Signed-off-by: paul luse <paul.e.luse@intel.com>