lib/blobstore: add clear_method param when deleting an lvol
Instead of providing clear_method on lvol create (removed in
previous patches) we now provide it on delete as an optional
parameter for the RPC.

Still TODO: lvolstore use of clear_method on delete in next
patch. Will need to have the lvolstore delete RPC also
accept clear_method to apply on all lvols that it currently
contains as the setting is not saved anywhere. This would
also make it consistent with how lvol delete works wrt
clear_method.

Fixes issue #944

Change-Id: I8f78e2fa25a602cac44c5a9182aafc04e728ccd3
Signed-off-by: paul luse <paul.e.luse@intel.com>