lib/blobstore: retrieve clear_method xattr before clearing device
When an lvol is created, the underlying blob's clear_method is stored
as an xattr but not retrieved in the function spdk_bs_batch_clear_dev()
which is used in both lvol creation and lvol deletion. By updating
the method in the local context the value of BLOB_CLEAR_WITH_NONE will
be respected.

Fixes issue #944

Change-Id: I8f78e2fa25a602cac44c5a9182aafc04e728ccd3
Signed-off-by: paul luse <paul.e.luse@intel.com>