nvme_rdma: add simple disconnect handling.
If we get a disconnect event, for now, assume that the controller is
failed and disconnect. This at least informs us that something has gone
wrong with the connection and allows us to quit the application early.

There should be more robust handling of these events in the future, but
this is step one.

Also add a test to confirm the target disconnect behavior.

Change-Id: I5bd2c33c04c55c0a6367877ec4d5c27bb234ebd0
Signed-off-by: Seth Howell <seth.howell@intel.com>