module/raid: implement func pointers for raid related functtions
Next patch will rename the relevant functions so it's clear they're
RAID 0.  This patch simply takes those functions which include RAID
specific mapping and assign them to existing functions via func
pointers in the RAID struct.

Some of these may not appear to be neccessary just yet, ie a RAID
level specific func pointer called from within a RAID specific
function, but this is a step towards refactoring to get the abstractions
in palce that make the most sense for both RAID0 and RAID1E as well
as future RAID levels.

Change-Id: I8c7724d855937a9c1ca78cdb8ec500521f23b12d
Signed-off-by: paul luse <paul.e.luse@intel.com>