bdev: Avoid executing spdk_bdev_calculate_measured_queue_depth after removing bdev
If you use iostat for a bdev and then delete the bdev (the
splk_bdev_calculate_measured_queue_depth poller is not removed during this
process), This will cause the bdev wild pointer or null pointer to operate
and access in the spdk_bdev_calculate_measured_queue_depth function.

In order to solve this problem, you can cancel this poller in the process
of deleting bdev.

Signed-off-by: Jian Wang <wangjian161@huawei.com>
Change-Id: I30c0a1b5d6b8d71d409a40e68c0ba4803e05dd85