bdevperf: Use zero copy bdev I/O APIs for write I/O
Add a bdev_io pointer to struct bdevperf_task to save and use later.

To make verify and reset feature usable even when zcopy is used for
write I/O, copy the to-be-verified buffer to the write buffer
allocated by zcopy API.

A few erroneous return codes are ignored or asserted because
error processing for zero copy write I/O will be implemented later.

Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>
Change-Id: I1a70a527a285c3ac4f86ebdef051cadf4620a37d